arc: rework

+++ files/board/arc/overlayfs/opt/arc/arc.sh
@@ -60,7 +60,6 @@ BOOTWAIT="$(readConfigKey "arc.bootwait" "${USER_CONFIG_FILE}")"
 REMAP="$(readConfigKey "arc.remap" "${USER_CONFIG_FILE}")"
 KERNELLOAD="$(readConfigKey "arc.kernelload" "${USER_CONFIG_FILE}")"
 ODP="$(readConfigKey "arc.odp" "${USER_CONFIG_FILE}")"
-MAC1="$(readConfigKey "arc.mac1" "${USER_CONFIG_FILE}")"
 
 ###############################################################################
 # Mounts backtitle dynamically
@@ -200,7 +199,8 @@ function arcMenu() {
     writeConfigKey "arc.remap" "" "${USER_CONFIG_FILE}"
     writeConfigKey "arc.paturl" "" "${USER_CONFIG_FILE}"
     writeConfigKey "arc.pathash" "" "${USER_CONFIG_FILE}"
-    writeConfigKey "sn" "" "${USER_CONFIG_FILE}"
+    writeConfigKey "arc.sn" "" "${USER_CONFIG_FILE}"
+    writeConfigKey "arc.mac1" "" "${USER_CONFIG_FILE}"
     CONFDONE="$(readConfigKey "arc.confdone" "${USER_CONFIG_FILE}")"
     BUILDDONE="$(readConfigKey "arc.builddone" "${USER_CONFIG_FILE}")"
     if [ -f "${ORI_ZIMAGE_FILE}" ] || [ -f "${ORI_RDGZ_FILE}" ] || [ -f "${MOD_ZIMAGE_FILE}" ] || [ -f "${MOD_RDGZ_FILE}" ]; then
@@ -295,13 +295,13 @@ function arcsettings() {
       if [ ${resp} -eq 1 ]; then
         # read valid serial from file
         SN="$(readModelKey "${MODEL}" "arc.serial")"
-        writeConfigKey "sn" "${SN}" "${USER_CONFIG_FILE}"
+        writeConfigKey "arc.sn" "${SN}" "${USER_CONFIG_FILE}"
         writeConfigKey "extensions.cpuinfo" "" "${USER_CONFIG_FILE}"
         writeConfigKey "arc.patch" "true" "${USER_CONFIG_FILE}"
       elif [ ${resp} -eq 2 ]; then
         # Generate random serial
         SN="$(generateSerial "${MODEL}")"
-        writeConfigKey "sn" "${SN}" "${USER_CONFIG_FILE}"
+        writeConfigKey "arc.sn" "${SN}" "${USER_CONFIG_FILE}"
         writeConfigKey "arc.patch" "false" "${USER_CONFIG_FILE}"
         writeConfigKey "extensions.cpuinfo" "" "${USER_CONFIG_FILE}"
       fi
@@ -311,7 +311,7 @@ function arcsettings() {
   else
     # Generate random serial
     SN="$(generateSerial "${MODEL}")"
-    writeConfigKey "sn" "${SN}" "${USER_CONFIG_FILE}"
+    writeConfigKey "arc.sn" "${SN}" "${USER_CONFIG_FILE}"
     writeConfigKey "arc.patch" "false" "${USER_CONFIG_FILE}"
     writeConfigKey "extensions.cpuinfo" "" "${USER_CONFIG_FILE}"
   fi
@@ -365,7 +365,6 @@ function arcsettings() {
 ###############################################################################
 # Building Loader
 function make() {
-  clear
   MODEL="$(readConfigKey "model" "${USER_CONFIG_FILE}")"
   PLATFORM="$(readModelKey "${MODEL}" "platform")"
   PRODUCTVER="$(readConfigKey "productver" "${USER_CONFIG_FILE}")"
@@ -391,16 +390,31 @@ function make() {
       return 1
     fi
   done < <(readConfigMap "extensions" "${USER_CONFIG_FILE}")
+  # Update PAT Data
+  dialog --backtitle "$(backtitle)" --colors --title "Arc Build" \
+    --infobox "Get PAT Data ..." 0 0
+  idx=0
+  while [ ${idx} -le 3 ]; do # Loop 3 times, if successful, break
+    PAT_URL="$(curl -skL "https://www.synology.com/api/support/findDownloadInfo?lang=en-us&product=${MODEL/+/%2B}&major=${PRODUCTVER%%.*}&minor=${PRODUCTVER##*.}" | jq -r '.info.system.detail[0].items[0].files[0].url')"
+    PAT_HASH="$(curl -skL "https://www.synology.com/api/support/findDownloadInfo?lang=en-us&product=${MODEL/+/%2B}&major=${PRODUCTVER%%.*}&minor=${PRODUCTVER##*.}" | jq -r '.info.system.detail[0].items[0].files[0].checksum')"
+    PAT_URL=${PAT_URL%%\?*}
+    if [ -n "${PAT_URL}" ] && [ -n "${PAT_HASH}" ]; then
+      break
+    fi
+    idx=$((${idx} + 1))
+  done
+  [ -z "${PAT_URL}" ] || [ -z "${PAT_HASH}" ] && return
+  writeConfigKey "arc.paturl" "${PAT_URL}" "${USER_CONFIG_FILE}"
+  writeConfigKey "arc.pathash" "${PAT_HASH}" "${USER_CONFIG_FILE}"
   if [ ! -f "${ORI_ZIMAGE_FILE}" ] || [ ! -f "${ORI_RDGZ_FILE}" ]; then
-    # Clean old files
+    # Clean old Files
     rm -rf "${UNTAR_PAT_PATH}"
     rm -rf "${CACHE_PATH}/${MODEL}/${PRODUCTVER}"
-    # Check for existing files
+    # Check for existing Files
     mkdir -p "${CACHE_PATH}/${MODEL}/${PRODUCTVER}"
     DSM_FILE="${CACHE_PATH}/${MODEL}/${PRODUCTVER}/dsm.tar"
-    DSM_MODEL="$(echo "${MODEL}" | sed -e 's/+/%2B/g')"
-    # Get new files
-    DSM_LINK="${DSM_MODEL}/${PRODUCTVER}/dsm.tar"
+    # Get new Files
+    DSM_LINK="${MODEL/+/%2B}/${PRODUCTVER}/dsm.tar"
     DSM_URL="https://raw.githubusercontent.com/AuxXxilium/arc-dsm/main/files/${DSM_LINK}"
     STATUS=$(curl --insecure -s -w "%{http_code}" -L "${DSM_URL}" -o "${DSM_FILE}")
     if [ $? -ne 0 ] || [ ${STATUS} -ne 200 ]; then
@@ -412,20 +426,14 @@ function make() {
       if [ $? -ne 0 ] || [ ${STATUS} -ne 200 ]; then
         dialog --backtitle "$(backtitle)" --title "DSM Download" --aspect 18 \
         --msgbox "No DSM Image found!\nTry Syno Link." 0 0
-        # Grep Values
-        PAT_MODEL="$(echo "${MODEL}" | sed -e 's/+/%2B/g')"
-        PAT_MAJOR="$(echo "${PRODUCTVER}" | cut -b 1)"
-        PAT_MINOR="$(echo "${PRODUCTVER}" | cut -b 3)"
         # Grep PAT_URL
-        PAT_URL="$(curl -skL "https://www.synology.com/api/support/findDownloadInfo?lang=en-us&product=${PAT_MODEL}&major=${PAT_MAJOR}&minor=${PAT_MINOR}" | jq -r '.info.system.detail[0].items[0].files[0].url')"
-        PAT_URL="${PAT_URL%%\?*}"
         PAT_FILE="${MODEL}_${PRODUCTVER}.pat"
         PAT_PATH="${CACHE_PATH}/dl/${PAT_FILE}"
         mkdir -p "${CACHE_PATH}/dl"
         STATUS=$(curl -k -w "%{http_code}" -L "${PAT_URL}" -o "${PAT_PATH}" --progress-bar)
         if [ $? -ne 0 ] || [ ${STATUS} -ne 200 ]; then
           dialog --backtitle "$(backtitle)" --title "DSM Download" --aspect 18 \
-            --msgbox "No DSM Image found!\ Exit." 0 0
+            --msgbox "No DSM Image found!" 0 0
           rm -f "${PAT_PATH}"
           return 1
         fi
@@ -451,7 +459,7 @@ function make() {
             ;;
         esac
         if [ "${isencrypted}" = "yes" ]; then
-            # Uses the extractor to untar PAT file
+            # Uses the extractor to untar PAT File
             LD_LIBRARY_PATH="${EXTRACTOR_PATH}" "${EXTRACTOR_PATH}/${EXTRACTOR_BIN}" "${PAT_PATH}" "${UNTAR_PAT_PATH}"
         else
             # Untar PAT file
@@ -474,11 +482,20 @@ function make() {
     cp -f "${UNTAR_PAT_PATH}/GRUB_VER"        "${SLPART_PATH}"
     cp -f "${UNTAR_PAT_PATH}/zImage"          "${ORI_ZIMAGE_FILE}"
     cp -f "${UNTAR_PAT_PATH}/rd.gz"           "${ORI_RDGZ_FILE}"
+    # Check Pat Hash
+    PAT_HASH="$(readConfigKey "arc.pathash" "${USER_CONFIG_FILE}")"
+    PAT_HASH_PRE="$(cat ${UNTAR_PAT_PATH}/pat_hash)"
+    if [ "${PAT_HASH}" != "${PAT_HASH_PRE}" ]; then
+      dialog --backtitle "$(backtitle)" --title "Error" --aspect 18 \
+        --msgbox "PAT Files do not match! Abort!!!" 0 0
+      return 1
+    fi
   fi
   # Reset Bootcount if User rebuild DSM
   if [ ${BOOTCOUNT} -gt 0 ] || [ -z "${BOOTCOUNT}" ]; then
     writeConfigKey "arc.bootcount" "0" "${USER_CONFIG_FILE}"
   fi
+  clear
   livepatch
   if [ ${FAIL} -eq 1 ]; then
     echo "Patching DSM Files failed! Please stay patient for Update."
@@ -508,7 +525,7 @@ function make() {
     fi
   else
     dialog --backtitle "$(backtitle)" --title "Error" --aspect 18 \
-      --msgbox "Build failed!\nPlease check your Internetconnection and Diskspace!" 0 0
+      --msgbox "Build failed!\nPlease check your Connection and Diskspace!" 0 0
     return 1
   fi
 }
@@ -529,7 +546,7 @@ function editUserConfig() {
   OLDPRODUCTVER="${PRODUCTVER}"
   MODEL="$(readConfigKey "model" "${USER_CONFIG_FILE}")"
   PRODUCTVER="$(readConfigKey "productver" "${USER_CONFIG_FILE}")"
-  SN="$(readConfigKey "sn" "${USER_CONFIG_FILE}")"
+  SN="$(readConfigKey "arc.sn" "${USER_CONFIG_FILE}")"
   if [ "${MODEL}" != "${OLDMODEL}" ] || [ "${PRODUCTVER}" != "${OLDPRODUCTVER}" ]; then
     # Remove old files
     rm -f "${MOD_ZIMAGE_FILE}"
@@ -780,8 +797,8 @@ function cmdlineMenu() {
   echo "1 \"Add/edit a Cmdline item\""                          >"${TMP_PATH}/menu"
   echo "2 \"Delete Cmdline item(s)\""                           >>"${TMP_PATH}/menu"
   echo "3 \"Define a serial number\""                           >>"${TMP_PATH}/menu"
-  echo "4 \"Add experimental CPU Fix\""                         >>"${TMP_PATH}/menu"
-  echo "5 \"Add experimental RAM Fix\""                         >>"${TMP_PATH}/menu"
+  echo "4 \"Add CPU Fix\""                                      >>"${TMP_PATH}/menu"
+  echo "5 \"Add RAM Fix\""                                      >>"${TMP_PATH}/menu"
   echo "6 \"Show user Cmdline\""                                >>"${TMP_PATH}/menu"
   echo "7 \"Show Model/Build Cmdline\""                         >>"${TMP_PATH}/menu"
   # Loop menu
@@ -848,7 +865,7 @@ function cmdlineMenu() {
           [ $? -eq 0 ] && break
         done
         SN="${SERIAL}"
-        writeConfigKey "sn" "${SN}" "${USER_CONFIG_FILE}"
+        writeConfigKey "arc.sn" "${SN}" "${USER_CONFIG_FILE}"
         writeConfigKey "arc.builddone" "false" "${USER_CONFIG_FILE}"
         BUILDDONE="$(readConfigKey "arc.builddone" "${USER_CONFIG_FILE}")"
         ;;
@@ -1230,7 +1247,7 @@ function backupMenu() {
                     if [ -n "${SN}" ]; then
                       deleteConfigKey "arc.patch" "${USER_CONFIG_FILE}"
                       SNARC="$(readConfigKey "arc.serial" "${MODEL_CONFIG_PATH}/${MODEL}.yml")"
-                      writeConfigKey "sn" "${SN}" "${USER_CONFIG_FILE}"
+                      writeConfigKey "arc.sn" "${SN}" "${USER_CONFIG_FILE}"
                       TEXT+="\nSerial: ${SN}"
                       if [ "${SN}" = "${SNARC}" ]; then
                         writeConfigKey "arc.patch" "true" "${USER_CONFIG_FILE}"
@@ -1395,7 +1412,7 @@ function backupMenu() {
                     if [ -n "${SN}" ]; then
                       deleteConfigKey "arc.patch" "${USER_CONFIG_FILE}"
                       SNARC="$(readConfigKey "arc.serial" "${MODEL_CONFIG_PATH}/${MODEL}.yml")"
-                      writeConfigKey "sn" "${SN}" "${USER_CONFIG_FILE}"
+                      writeConfigKey "arc.sn" "${SN}" "${USER_CONFIG_FILE}"
                       TEXT+="\nSerial: ${SN}"
                       if [ "${SN}" = "${SNARC}" ]; then
                         writeConfigKey "arc.patch" "true" "${USER_CONFIG_FILE}"
@@ -1929,7 +1946,7 @@ function sysinfo() {
       fi
       NETIP=$(ip route show dev ${ETHX[${N}]} 2>/dev/null | sed -n 's/.* via .* src \(.*\)  metric .*/\1/p')
       ARCIP="$(readConfigKey "arc.ip" "${USER_CONFIG_FILE}")"
-      if [ "${ETHX[${N}]}" = "eth0" ] && [ "${ARCIP}" != "" ] && [ ${BOOTCOUNT} -gt 0 ]; then
+      if [ "${ETHX[${N}]}" = "eth0" ] && [ -n "${ARCIP}" ] && [ ${BOOTCOUNT} -gt 0 ]; then
         NETIP="${ARCIP}"
         MSG="STATIC"
       else
@@ -1937,7 +1954,7 @@ function sysinfo() {
       fi
       if [ -n "${NETIP}" ]; then
         SPEED=$(ethtool ${ETHX[${N}]} | grep "Speed:" | awk '{print $2}')
-        TEXT+="\n  ${DRIVER} (${SPEED} | ${MSG}): \ZbIP: ${NETIP} | Mac: ${MAC}\Zn"
+        TEXT+="\n  ${DRIVER} (${SPEED} | ${MSG}) \ZbIP: ${NETIP} | Mac: ${MAC}\Zn"
         break
       fi
       COUNT=$((${COUNT} + 1))
@@ -2297,12 +2314,10 @@ function boot() {
   if [ $? -eq 0 ]; then
     make
   fi
-  if [ "${DIRECTBOOT}" = "false" ]; then
-    grub-editenv "${GRUB_PATH}/grubenv" create
-  fi
   dialog --backtitle "$(backtitle)" --title "Arc Boot" \
     --infobox "Booting to DSM - Please stay patient!" 0 0
-  exec reboot
+  sleep 2
+  boot.sh
 }
 
 ###############################################################################
@@ -2448,6 +2463,7 @@ while true; do
     p) bootipwaittime; NEXT="p" ;;
     -) bootwaittime; NEXT="-" ;;
     q) [ "${DIRECTBOOT}" = "false" ] && DIRECTBOOT='true' || DIRECTBOOT='false'
+      grub-editenv "${GRUB_PATH}/grubenv" create
       writeConfigKey "arc.directboot" "${DIRECTBOOT}" "${USER_CONFIG_FILE}"
       writeConfigKey "arc.bootcount" "0" "${USER_CONFIG_FILE}"
       NEXT="q"
